requirements.txt
click==7.1.2
flake8==3.8.4
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
forbiddenfruit==0.1.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==8.0.1
pycodestyle==2.6.0
pyflakes==2.2.0
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
